+++ cutile-examples/examples/async_saxpy.rs
@@ -24,7 +24,7 @@ mod my_module {
         x: &Tensor<T, { [-1, -1] }>,
     ) {
         let tile_a = a.broadcast(y.shape());
-        let tile_x = load_tile_like(x, y);
+        let tile_x = x.load_like(y);
         let tile_y = y.load();
         y.store(tile_a * tile_x + tile_y);
     }